+++ media/js/src/JXGBoard.jsx
@@ -4,7 +4,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {MathComponent} from 'mathjax-react';
 import JXG from 'jsxgraph';
-import {graphTypes, mkDemandSupply, mkNonLinearDemandSupply} from './Graph.js';
+import {graphTypes} from './graphs/graphTypes.js';
+import {
+    mkDemandSupply, mkNonLinearDemandSupply
+} from './graphs/Graph.js';
 import AreaDisplay from './AreaDisplay.jsx';
 import {getL1SubmissionOffset, getL2SubmissionOffset} from './utils.js';
